feat:add validate address form and add service on home page
