+++ package.json
@@ -26,6 +26,7 @@
     "date-fns": "^3.3.1",
     "dayjs": "^1.11.10",
     "framer-motion": "^11.0.3",
+    "joi": "^17.12.2",
     "js-cookie": "^3.0.5",
     "material-ui-popup-state": "^5.0.10",
     "next": "^14.1.0",
